+++ service/core/action/tag/delete.go
@@ -53,6 +53,15 @@ func delete(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	err = config.DB.Where(&model.PostTag{
+		TagID: uint(id),
+	}).First(&model.PostTag{}).Error
+
+	if err == nil {
+		errorx.Render(w, errorx.Parser(util.CannotDeleteError()))
+		return
+	}
+
 	config.DB.Delete(&result)
 
 	renderx.JSON(w, http.StatusOK, nil)
